Fixed packaging issues.
Removed packages from requirements file which are installed from depencies.

Added new files for packages which need to be installed without depencies.
django-filer has depencies which are clashing with the packages that are part
of the django-cms package.

django-publisher problems:

- django-publisher has been extended by the django-cms team. So it's not
possible to use the package from PyPI.
- Removed publisher from 2.2 settings
(it's now packaged as a module in django-cms).